Build Intcode computer internal architecture
A "computer" running on integer codes.

Continue development of the Intcode computer for AoC 2019.

Develop the implementation of the internal architecture for the AsmComp
computer. Roughly, very roughly, based on the 6502C processor, which I
was familiar with a few decades ago. I don't envision the majority of
what I create being needed, but it is a starting point.

Key points of difference number range implementation and flag usage. I
have also modified the register set some. The 6502C is an 8-bit CPU, and
the carry/overflow flags reflect that in their native implementation. I
will not restrict integers to 8 bits, using instead the native numbers
from the Perl I have, 64-bit limit. As an additional consequence of that
I will not implement the carry and overflow flags, and I don't expect to
have any issues from the challenges for lacking them.

The choice to modify the registers was driven, again, by expectations of
the challenges to be seen. I've added a register `D` to use as the other
number in math operations, applied to the accumulator rather than using
the referenced memory, or operand. This is an attempt to make the built
computer easier to mold into whatever is required for the challenges. In
addition I added a register `I` for indexed memory access. The 6502C has
both an `X` and `Y` register which can be used as the base for indexing,
while also being used for other general purposes. I've kept the `X` for
general use, while moving the indexing to the new `I` and dropping the
`Y` register completely. Lastly, in regard to the registers, they have
all been converted to single character names:

*   `PC`->``, Program counter (Code pointer)
*   `AC`->``, Accumulator
*   `SR`->``, Status register
*   `SP`->`F`, Stack pointer (Flags)

I have also made signifcant change to the addressing modes. For starters
I have completely ignored the concept of zero-page addressing vs normal
addressing. Two reasons; one, the 6502C used zero-page mode as a cycle-
saving mechanism, and two, I don't envision any benefit in connection to
the AoC challenges for having the difference. In reworking the registers
by removing the `Y` I also removed it from use in addressing, and with
the switch to the `I` register for address indexing, the `X` is also not
used. I've choosen to have five modes: immediate, absolute, indirect,
indexed, and indirect indexed.

*   Immediate considers the operand as the
data, without addressing of any kind. Use of the accumulator in some
cases will replace immediate mode, and for op codes without data, I'll
probably consider it as immediate as well.
*   Absolute mode takes the operand as the address of the data, with no
futher translation needed. `memory[op]`
*   Indirect mode is, in essence, a pointer. The operand is the address
in which is found the address of the data. `memory[memory[op]]`
*   Indexed mode will use the `I` register as the base address, which is
then modified by the operand to find the actual address of the data. In
other words, the operand acts as the index into an array with the zeroth
element at the address in the `I` register. Signed operands will access
memory before the base, rather than as an index from the end of the list
as is done in some languages. `memory[I+op]`
*   Indirect indexed is the tricky one. It can go three ways:
`memory[memory[I+op]]` or `memory[memory[I]+op]` or `memory[I+memory[op]]`
all of which, given the same set of values for `I`, operand, and memory
can result in completely different final addresses. My choice, for the
sake of having a standard, is to use the first option. It seems the most
obvious to me. It's likely to make no difference anyway as it seems not
very likely that the challenges will involve such convoluted addressing.

After reducing the versions of addressing available, I intend to enhance
the addressing available per op code. Many of the op codes for the 6502C
are quite limited in the available addressing. Most likely as a result
of the CPU design limitations. Where possible I want to add all address
modes and as many as possible where limits seem to be involved. One case
I cannot change is having Immediate mode available for reading from any
of the registers. That's just nonsensical.

As the up-chain modules are developed, there's likely to be further work
needed here. Noticably abscent is any method for accessing or changing
any of the flags, registers, or memory. Any use of the stack pointer is
likewise probably going to be ignored. Unless it is somehow involved in
on of the challenges, there's no need to deal with it here. Just let the
modules use Perl's stack natively and ignore what would be memory issues
in the 6502C.

Developmental note:

Contrary to normal development, this is being built from the bottom up,
rather than in my normal Dijkstra-isc style of top down decomposition.
Firstly, this is a result of my "seeing" the solution as a whole system
rather than working my way through the process. Secondly, having had the
epiphany, it seemed like a good oportunity to try the details-out method
as an experience (probably not to be repeated). With that in mind, I'll
have to be careful to have all modules supply hooks to their parents.

Side note:
Writing this took longer than the code it is attached to.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>