Add low-level memory routines.
Two routines, `base_read_memory` and `base_write_memory` are internal
routines to access the memory using the address mode supplied in the
arguments.

The `system_memory` is the interface exposed to the using module which
abstracts the memory access. Of special note are the calling options.:

* `system_memory()` [no args] results in a raw memory dump for program
ending, diagnostic, other purposes, and returns false.
* `system_memory(1,15)` [2 args] reads, and returns the contents, of
the memory referenced by the second argument, potentially modified in
accordance to the address mode indicated by teh first arg. The first
argument should be supplied using the given address mode variable, not
using hard-coded numbers.
* `system_memory(0,43,4569)` [3 args] writes the 3rd argument to memory
using the same concept of address mode used for reading memory above.
The return value is the number written to memory.
* `system_memory(1,2,3,..,45)` [more than 3 args] copies the list of
arguments directly into memory, starting at address 0, and returns the
size of the used memory.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>