Implement the stack handling for the computer
Following the same process as before, the exported routinte is a
wrapper for the access routines.

There is no provisions for accessing arbitrary elements of the stack.
Either it is pushed or pulled. The stack pointer can, however, be
manipulated using the system_register routine, to effectively work
around that limitation. The stack will grow as needed with calls which
push onto the stack, it is not, however, reduced in size when a pull is
done. Values placed in the stack remain there until a new value is put
in its place. A dump of the stack will list all data in the stack with
no respect to the current stack pointer.

While possible, it shouldn't be done where the stack is manipulated by
way of adjusting the stack pointer.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>