Add opcode handling
Create a set of opcodes, similar to the 6502 chip instruction set, to
implement in the computer. The adaptation is loose at best. Much of the
changes, and the reasoning, were already discussed.

Error checking has been mostly created to fail silently. The "programs"
provided by the AoC challenges should be operational, and the API, from
Elf to Asm should be sufficient to handle everything.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>