Internalize the running of the code.
Switch from the use of the opcodes as the interface to a list of codes
to run as the interface. Now the commands are passed in as a list of
values, the code to run and when needed the access mode and operand. The
module handles interating through the list, including knowing how many
of the list elements are used for each instruction.

The only exported symbols are now the routine to run the commands and a
routine for loading the initial state of the memory.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>