Name change and cleanup
Changed the concept of 'op codes' to 'assembly codes' as I am using the
mnemonic codes rather than numeric codes. Names of symbols and modules
are changed to agree with that concept switch.

Added STP and ABT codes, and modified BRK code, to provide a variety of
terminal conditions. In all cases the machine state is dumped to the CLI
and the program is likely to halt. BRK will end the running of the given
snippet, while the Perl program will continue. STP and ABT (stop and
abort) will exit the Perl program completely. Futher knowledge of the
requirements for the Elves may dictate a change, or completely ignore
both options.

Walked through the code and made routine calls follow the standard
pattern of `sub_name(arg1, arg2)` even though parens are optional in
most cases within Perl. Makes it easier for others to read and know how
the code should behave.

Fixed an oversight in base_read_memory of BaseCode.pm.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>