Adjust the assembly computer to be operational
Thus far the AsmComp system has been built to be a backend, or helper,
for the elves computer. I'm so far behind the dates on the AoC for 2019
that I've decided to enhance the AsmComp to be operational as a stand-
alone as well.

Added a code segment to hold the program, independent from the data. I,
of course, have drifted even farther from the 6502 model by borrowing a
concept from the x86 CPU family. The division between code and data is,
in this case absolute. The program will have no way to modify itself.
Addresses for branching and jumping can be held in memory, and basically
mirror the addressing modes of data, so complex dispatch tables can be
built.

I had to make the code and data isolated segments since the AsmComp is
not going to be a one-to-one translation of op codes from the elves
computer. As I have no idea if the elves computer will be doing any form
of branching, I cannot safely use the data memory for storing Asm code,
even though it will be used for storing elf code.

I will still keep the operations of each module encapsulated, with the
interface for the next higher module being passed through the EXPORT
functionality.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>