Implement day 1 part 1 using the AsmComp
I used Perl to generate the program code rather than doing it as an
external text file. The only significant difference between the two
methods is using `=>` instead of a comma to avoid needing to quote
all the non-numerics; Asm codes and address modes, and using variables
in Perl to hold the memory address information. Using the variables
makes it easier to see what is being referenced by the address than a
bald memory address would.

Also corrected an error in the relative address mode for code segment
addresses.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>