Remove need for environment variable
The newer versions of Perl have removed the current directory from the
`@INC` list unless the environment variable `PERL_USE_UNSAFE_INC` is
set to 1. A good practice for normal operations. Not so handy for such
little projects as this challenge will be. See the README file for
discussion of this more.

Updated the README and Changes files as well.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>