Modify the stopping operations
The assembly codes `BRK` (break), `ABT` (abort), and `STP` (stop) are
set to behave in different manners. Hopefully providing a useful
interface to the calling modules.

`ABT` merely dumps the core (contents of memory, registers and flags),
and does a hard exit. Presumably the exit is not preventable by the
rest of the "system" built. Divide by zero might be such a place to use
`ABT`. None of the flags are modified, as the system will be stopping
anyway.

`STP` sets the Break flag and provides a core dump and returns nothing.
The expectation is that `STP` is the end of the program, and nothing
more will be done with the existing system. Programs designed to run on
the AsmComp should use `STP` as the last assembly instruction. The
coding for the routine will terminate if no more memory locations are
loaded, of course, but the `STP` will provide a cleaner exit, and will
return a vaule to the caller to indicate it has officially stopped.

`BRK` stores the instruction pointer and current status register value
on the stack and then sets the Interrupt flag before returning. This
allows the program to resume at the next instruction if called to do so
by the higher module. (Routine to handle that is yet to be written.) A
different return value from the running program will alert the calling
module to the interrupt status rather than program termination.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>