Implement the resume option for programs
Add the `program_resume` rountine to handle program rentry after a
`BRK` instruction is encountered.

Added in the return codes to inform the calling module of how the code
was terminated. True (1) is normal program `STP`, False (0) is from a
`BRK` instruction, and `undef` (also false) indicates that there was no
instruction to terminate, just that the end of the instruction was
reached and there was nothing more to do.

Of note is that the `program_resume` routine includes the `RTI` code,
and such should not be within the program itself. The concept here is
that the running program includes a `BRK` to allow the higher module to
do something based on the data in the core dump, which is not part of
the operations of the AsmComp itself. This could, for example, be the
output of information to the terminal, of waiting for the userland to
complete some other operation. Anything which is not what the program
within the AsmComp is expected to do. After such actions are complete,
the supervising program can cause the program to resume. The `RTI` in
the routine will allow the AsmComp to reload the data it should and
continue from the instruction following the `BRK`.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>