Add routine half way between run and resume
The routine `program_step` if for running the program in little pieces,
or snippets. Each snippet acts like its own program, staring from the
code index of zero, but the flags and other registers are retained from
one invocation to the next. Each snippet should end with `BRK` or the
whole will fail. The provided exception is that `program_step` may be
invoked for a first snippet where there was no prior break. This is to
be detected by finding the stack empty, in which case the program will
silently switch to using the `program_run` routine.

Propogate the new routines to the AsmComp file and mark same as being
exported to the code using the AsmComp.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>