Add and export routines for elf computer access
Some of what will be needed by the ElfComp is along the lines of what
would be available to an operator using a breadboarded CPU. Things
such as direct editing/viewing of memory, which is not normally open
to the operator of a computer, but is typically built in to a bread-
board environment. Read/write memory, increment/decrement pointers, and
single-step programs come to mind at the moment.

The AsmCodes file has `direct_memory_access` which is propogated up to
be made available to the AsmComp API. Address mode is always absolute
in this case, and is hard-coded as such. ElfComp will need to read the
next memory location, as well as track its own position within the
memory, to create the assembly code snippets to hand off to the AsmComp
for execution. Writing may not be needed. Registers and flags are not
added, but may be if the need is seen.

Also converted routine propagation from routines implementing routines
to namespace aliasing. It may not matter at the level of use expected
for the challenges. Nevertheless, it's probably a good idea to do.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>