Modify stop conditions for cleaner output
I'm already tired of the core dump, and don't want to have it show up
every time. It may be needed, either for testing or for some future
challenge, so I won't remove it. Instead what I've done is repurpose
the Decimal flag. As the challenge involves an _Integer_ computer, I
don't expect that the Decimal flag will have any purpose. The flag now
will control the core dump: set is dump, clear is no dump, for both of
the `BRK` and `STP` codes. The `ABT` still does a core dump. It is an
**ABORT** after all. The expectation here is that in some rare case of
the core dump being needed by the challenge, the API can be used to set
the Decimal flag (by executing a snippet containing a single `SED`).

The new ElfComp command `enable_dump` when called with an argument of 1
will enable the core dump. Any other argument will turn off the dump.
Turning off the dump should not be needed, as it's off by default. It
can be used in some case where there are multiple BRK commands and one
needs to have the dump and the other do not. Not expected, but still it
is able to be done.

The routine `enable_dump` handles this for the ElfComp;

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>