Move solutions to directory with universal launcher
The solutions already created have been moved to the Solutions directory
and a new file `AoC-2019.pl` created to handle dispatching to the proper
solution files. Simple CLI interface of `AoC-2019.pl <day> <version>`.
This will require that all data files are in the Data directory with a
name of the format AoC-2019-dd.txt, where `dd` is the two-digit day for
the challenge. The `<version>` is either 'a' or 'b' for part 1 of the
challenges or b for part 2.

I also modified the solution for day 1 to preserve both versions of the
solution, with a seperate file for each version, and the supporting
routines doing the correct summation.

Did a little clean up in the process. Added the final `1;` to some of
the files, and removed Data::Dumper includes, where found.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>