Make computer re-entrant.
The flags, registers and memory of the AsmComp and ElfComp were set
and controlled at program start, and were one-shot style. The next part
of the challenges is easier if I can do repeated invocations of the
computer without restarting the Perl script. There is now the ability
to do a warm boot, or soft boot, reset of the AsmComp and ElfComp.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>