Day 4 part 1 solution
I can see the sequence of sums here, but coding it is too much trouble.
A simple brute force approach is so easy to write that it's the method
I'm going to use here. Part 2 of the challenge might make me wish I'd
gone the mathematical route, but who knows.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>