Increase output options for the AsmComp
Added the ability to output in binary `OPB`, octal `OPO`, hex `OPX`,
and bare decimal `OPD`. The comprehensive output `OUT` is probably not
that useful to "programs" that use the AsmComp, and `OPD` is what I am
presumming will be the standard used. I am also thinking that there may
come a time when the output is piped and controlling, minimizing, the
output will help in that regard.

Since the instructions for day 5 only mention output of numbers, the
new Elf opcode of 4 will be linked to the `OPD` AsmCode. I am going to
leave the use of ARGV for the input and have the ElfComp request the
input from the user and place it in ARGV for the AsmComp to read.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>