Change version number system
Rather than using numbers in the range of 0.01.01 - 0.01.25, as I had
begun with, I've switched to 0.19.01 - 0.19.25. The 01 - 25 indicates
the Day associated with the creation, or update, of the file. The 19 is
for 2019, the year of the challenges. Previous challenges go back to
2015, and those would be 0.15.01 - 0.15.25.

All files have been changed, including the Changes file, and a note has
been included in the README.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>