Enhance AsmCodes for X flag and D register
The Day 5 part 2 challenge exposed a gap in the Asm to Elf API. There
is no connection between a branch or jump in ElfCodes to the same in
AsmCodes. The AsmComp uses its own code pointer and has no access to
the ElfComp code pointer. With the lack of a one-to-one relation from
the memory of the computers and the AsmCodes as executed, there was no
method for adjusting the ElfComp code pointer based on an action in the
AsmComp.

The solution is to use the AsmComp's unsused flag bit, X, as a signal
to the API to return a value. The value to be exposed is in the AsmComp
D register. The D register was created, and used, as an intermediary
already, so it was the logical choice. The interactions might be tricky
to work out, but should be achievable.

To facilitate this, AsmCodes were added to access both the D register
and the X flag, along with some adjustments to the Txx inter-register
transfer code operations to seem more balanced and logical to me than
was listed in the 6502 data.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>