Change I/O procedures for AsmComp and ElfComp
I've decided that the Perl variable `ARGV` can be utilized as a channel
of communication between the AsmComp and ElfComp, as well as between
the "computer" and the user. I want the "normal" AsmComp I/O to use
`ARGV` which can be controlled, and read, by the supervising Perl, as
well as by the ElfComp when it is acting as a frontend for the AsmComp
backend. From a Perl POV the values will be treated as strings, while
both of the computers will consider them to be numbers, for input, and
as either numbers or strings for output, as appropriate for the AsmCode
or ElfCode.

The existing output functions will be converted to place their "output"
as the first argument to `ARGV`, Perl `unshift`. The input function is
already set to read the first argument using `shift`. Two new AsmCodes
will be created, `PLV` (pull value), and `PHV` (push value), which will
do Perl `pop` and `push` operations to `ARGV`. To maintain the ability
to actually output to the user terminal without intervention, a new
AsmCode `PRT` (print) will be utilized. Finally, a pair of new input
operations will be created. `REA` (read) will do a direct terminal
prompt and read, and `GET` will do a blocking read from `ARGV`. I see
no purpose for the latter, unless concurrent programming becomes an
issue.

These changes will also require back-porting to some of the completed
projects. Which ones are affected will be resolved by direct testing.

Updates to cover neglected records in the Changes file and to add a
notice of this change in the README.md file.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>