Create module for permutations
There are two versions, recursive and non-recursed. The recursed
version expects the routine to handle each iteration to be supplied in
the subroutine call, and is dispatched as permutations are found.

The interative version compiles and returns the complete list.

In both cases, the permutation is an array_ref, and is processed the
same way outside of the permutation generator.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>