Change concept to API, over computer access
Make a change so that the ElfComp system accesses the AsmComp as an API
rather than through the AsmComp file. Of course, the file is nearly the
same as it was, just moved. The difference will be wider as I adjust
other elements to work with the API or the computer.

The extra-credit solutions for day 1, for example, are built to use the
AsmComp as a computer. The solutions for day 2 use the ElfComp, which
was connected to the AsmComp module. Now that connection will be to the
API module. There are, or will be, things which the API can handle that
cannot be done using the AsmComp module, and other things which cannot
be accomplished using the API yet are easily done with the AsmComp file
such as reading a file into the AsmComp memory.

The plan here is for I/O for the Asm computer, when used directly to be
handled by the AsmComp module, while I/O for the Elf computer which is
connected to the Asm computer as a backend, will be handled by the API
and pass through the Elf computer before connecting with the user. This
will require some back-porting of changes to the completed solutions as
well as changes to the modules themselves.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>