Make I/O adjustments for assembly codes
Converted the output codes to use `push` rather than `unshift`. This
keeps the command line arguments, if any, intact, and allows the next
level up to access both the existing arguments and the return value in
the same place. Use `shift` to get the next command line argument, and
`pop` to get the last return value, with the ability to dig into the
data even farther using multiple shift or pop commands.

Working through the assembly codes I decided to document them, since I
have deviated so far from the prototype of the 6502 chip's assembly
codes. Doing so exposed a few places where I could add more codes for
balance, and helped me to see a better way to communicate with the
supervising code on "program" exit.

In essence, the terminal conditions are recorded in three flags, and
their binary sum is the return code from launch of the program. The
requested data is returned in `ARGV` as everything else is now. The
input and output from `STDIN` can now be faked, or piped, as can the
output for `STDOUT`.

Along the same lines, the routine names have been modified in a few
cases. It helps to organize them, helps find breaking points in the
calling modules, when I get to that, and helps to keep things in an
arrangement between levels for lightened cognitive load.

It is my hope that the operations of the AsmComp is done at this level
and that the API and user interface can become the focus of development
for the AsmComp.

The ElfComp, and existing solutions still need testing, and possibly
updates, to match the changes here.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>