Add simulated terminal interface for AsmComp
Since the use of the AsmComp as a backend now uses the API concept, I
need a module to represent the AsmComp on its own, as a stand-alone,
not the backend for something else.

Keeping things simple, it is nothing more than a place to supply the
data for the code segment and data segment, and possibly the "input"
values for ARGV, and launch the "program."

The ability to restart the program, and to resume the program after a
BRK command as well as the ability to directly access memory locations
is present as well. The memory access and the continue option are more
for debugging of the assembly code than for programmatic use.

Some minor changes have been made to the supporting modules, including
splitting the soft_start functionality between soft_start and
hard_start loading, loading, or not, the program code.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>