Update the ElfCodes to match the API
The ElfCodes module uses older style of interface to the the AsmComp.
The change to use the API requires updates to the calling modules.

Took this opportunity to update, add, documentation to all of the
modules in the IntComp directory. A few routines have been reordered
and some cleanup of comments, now in the docs. In general, it is my
hope that updates and changes to the modules should be rather small
from this point forward. Still 18 days of challenges to go, so it may
be a futile hope.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>