Update existing solutions for current AsmComp
The existing solutions have been updated to reflect the changes in the
"computers" behind them. Added a "version" string to all solutions,
including solutions which are not tied to the computers, as a standard
to use going forward. The version, as in the supporting modules, is to
reflect the "Day" involved with the updates. The solutions which are
not affected by the computer updates kept their day number, the others
have been updated to day 7.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>