Add output controls to the ElfComp
Day 7 challenge will involve many repeated runs of the program, five at
a shot. I'm not going to be sitting at the keyboard entering in each of
the permutated combinations of settings, and the stage outputs into the
next stage for each of the permutations. That's what computers are for.

Rather than make further updates to existing solutions, I'm adding a
new feature, default behavior is backward compatible, to redirect the
"output" to memory and allow the calling module to retrieve it.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>