Add file handle abilities to the AsmComp
Reading the day 7 instructions closer, I realized that it's meant to be
set so that all five "amplifiers" are running at the same time, not in
series, as was the plan. I decided to do simple `fork` calls since I
doubt that the challenges will reach the level where I'd need to use
something as full as [POE](https://metacpan.org/pod/POE).

For completeness sake, I'm adding `STDIN`, `STDOUT`, and `STDERR` to
the apropriate places to make commandline redirection work sensibly as
well. From the point of view the affected commands will use a supplied
file handle, or default to normal if none is given.

Only the I/O intended for API use will be affected: `INP` `OTB` `OTD`
`OTH` and `OTO`. The remaining I/O is for either debugging, or the
simulated termianl use, and they will only be subject to commandline
redirection.

A new hash `%io_handles`, with members `input` and `output` will hold
the redirection file handles for use. The AsmCodes module does most
of the work with `%io_handles`. The API propogates the hash for other
modules to access as `%asm_handles`.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>