Extend the new I/O direction into the ElfComp
With the I/O redirection enabled in the AsmComp using the API, it's
time to propogate the changes into the ElfComp. Most of the work is
done by the ElfCodes module. Added to ElfCodes is the routine to set
the AsmComp's file handle needs to the ones created for the challenges.

The ElfComp propogates `set_pipes` as `asm_pipes`. The routine will set
the supplied handles in the AsmComp API `%io_handles` hash. The calling
modules are responsible for opening the pipes and assigning them their
handles, as well as knowing which handle serves what purpose.

The Changes file is updated with this modification as well, and the
README reflets this enhancement.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>