Change to single files daily
Names on files are changed to "day_??" rather than "AoC-2019_??". The name of
the repo and the name of the directory ought to be enough to indicate what the
files are for.

The launcher is also renamed, and logic changed to handle other options. Options
include having alternative solutions. Day 1 has a secondary solution using the
IntCode Computer, and day 2 has an alternative where the brute force approach
was created for timing comparisons. (Wow, big difference!).

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>